made cli program
