+++ src/pedsilicoICH/__init__.py
@@ -3,4 +3,4 @@ from . import image_acquisition
 from . import lesion_definition
 from . import ground_truth_definition
 from . import artifact_generation
-from . import pipeline
+from . import study
